implemented noise detection model (tested and works)

+++ backend/schemas.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 class UploadResponse(BaseModel):
     status: str
-    confidence_score: int
+    confidence_score: float
     mask_url: str
     org_url: str
     coords: list
